lift up click handler to search bar
