+++ src/components/search-bar/search-result.tsx
@@ -1,18 +1,23 @@
 import { EntityModel } from '@/models/entity-model'
 import { EquipmentWithVariants } from '@/models/equipment'
-import { Rarity } from '../card/traits-list'
-import { ReactNode } from 'react'
-import React from 'react'
 import { useRouter } from 'next/navigation'
+import { ReactNode } from 'react'
+import { Rarity } from '../card/traits-list'
 
-export function SearchResult({ data }: { data: EntityModel }) {
+export function SearchResult({
+  data,
+  handleClick,
+}: {
+  data: EntityModel
+  handleClick: () => any
+}) {
   const router = useRouter()
 
   const level: string | undefined = buildLevelDisplay(data)
   return (
     <div
-      className="flex justify-between cursor-pointer"
-      onClick={() => router.push(`/reference/${getRoute(data)}/${data._id}`)}
+      className="p-1 border-x border-t border-stone-300 hover:bg-stone-700 flex justify-between cursor-pointer"
+      onClick={handleClick}
     >
       <span>
         {data.name} {buildRarityDisplay(data)}
@@ -22,20 +27,6 @@ export function SearchResult({ data }: { data: EntityModel }) {
       </span>
     </div>
   )
-}
-
-function getRoute(data: EntityModel): string {
-  switch (data.entity_type) {
-    case 'EQUIPMENT':
-    case 'EQUIPMENT_WITH_VARIANTS':
-      return 'equipment'
-    case 'SPELL':
-      return 'spells'
-    case 'CONDITION':
-      return 'conditions'
-    case 'TRAIT':
-      return 'traits'
-  }
 }
 
 function getType(data: EntityModel): string {
